Removed globals.py entirely
